Reorganized change-point tests
- move `ref_changepoints()` to reference.py
- move changepoint tests to test_changepoints.py
